split build graph target metadata tests
